Modularize Ocarina Time Travel hook
