+++ soh/soh/Enhancements/mods.cpp
@@ -59,101 +59,6 @@ static const ALIGN_ASSET(2) char tokinoma_room_0DL_007A70[] = dtokinoma_room_0DL
 #define dtokinoma_room_0DL_007FD0 "__OTR__scenes/shared/tokinoma_scene/tokinoma_room_0DL_007FD0"
 static const ALIGN_ASSET(2) char tokinoma_room_0DL_007FD0[] = dtokinoma_room_0DL_007FD0;
 
-/// Switches Link's age and respawns him at the last entrance he entered.
-void SwitchAge() {
-    if (gPlayState == NULL)
-        return;
-
-    Player* player = GET_PLAYER(gPlayState);
-
-    // Hyrule Castle: Very likely to fall through floor, so we force a specific entrance
-    if (gPlayState->sceneNum == SCENE_HYRULE_CASTLE || gPlayState->sceneNum == SCENE_OUTSIDE_GANONS_CASTLE) {
-        gPlayState->nextEntranceIndex = ENTR_CASTLE_GROUNDS_SOUTH_EXIT;
-    } else {
-        gSaveContext.respawnFlag = 1;
-        gPlayState->nextEntranceIndex = gSaveContext.entranceIndex;
-
-        // Preserve the player's position and orientation
-        gSaveContext.respawn[RESPAWN_MODE_DOWN].entranceIndex = gPlayState->nextEntranceIndex;
-        gSaveContext.respawn[RESPAWN_MODE_DOWN].roomIndex = gPlayState->roomCtx.curRoom.num;
-        gSaveContext.respawn[RESPAWN_MODE_DOWN].pos = player->actor.world.pos;
-        gSaveContext.respawn[RESPAWN_MODE_DOWN].yaw = player->actor.shape.rot.y;
-
-        if (gPlayState->roomCtx.curRoom.behaviorType2 < 4) {
-            gSaveContext.respawn[RESPAWN_MODE_DOWN].playerParams = 0x0DFF;
-        } else {
-            // Scenes with static backgrounds use a special camera we need to preserve
-            Camera* camera = GET_ACTIVE_CAM(gPlayState);
-            s16 camId = camera->camDataIdx;
-            gSaveContext.respawn[RESPAWN_MODE_DOWN].playerParams = 0x0D00 | camId;
-        }
-    }
-
-    gPlayState->transitionTrigger = TRANS_TRIGGER_START;
-    gPlayState->transitionType = TRANS_TYPE_INSTANT;
-    gSaveContext.nextTransitionType = TRANS_TYPE_FADE_BLACK_FAST;
-    gPlayState->linkAgeOnLoad ^= 1;
-
-    // Discover adult/child spawns
-    if (gPlayState->linkAgeOnLoad == LINK_AGE_ADULT) {
-        Entrance_SetEntranceDiscovered(ENTR_HYRULE_FIELD_10, false);
-    } else {
-        Entrance_SetEntranceDiscovered(ENTR_LINKS_HOUSE_CHILD_SPAWN, false);
-    }
-
-    static HOOK_ID hookId = 0;
-    hookId = REGISTER_VB_SHOULD(VB_INFLICT_VOID_DAMAGE, {
-        *should = false;
-        GameInteractor::Instance->UnregisterGameHookForID<GameInteractor::OnVanillaBehavior>(hookId);
-    });
-}
-
-/// Switches Link's age and respawns him at the last entrance he entered.
-void RegisterOcarinaTimeTravel() {
-
-    GameInteractor::Instance->RegisterGameHook<GameInteractor::OnOcarinaSongAction>([]() {
-        if (!GameInteractor::IsSaveLoaded(true) || !CVarGetInteger(CVAR_ENHANCEMENT("TimeTravel"), 0)) {
-            return;
-        }
-
-        Actor* player = &GET_PLAYER(gPlayState)->actor;
-        Actor* nearbyTimeBlockEmpty =
-            Actor_FindNearby(gPlayState, player, ACTOR_OBJ_WARP2BLOCK, ACTORCAT_ITEMACTION, 300.0f);
-        Actor* nearbyTimeBlock = Actor_FindNearby(gPlayState, player, ACTOR_OBJ_TIMEBLOCK, ACTORCAT_ITEMACTION, 300.0f);
-        Actor* nearbyOcarinaSpot = Actor_FindNearby(gPlayState, player, ACTOR_EN_OKARINA_TAG, ACTORCAT_PROP, 120.0f);
-        Actor* nearbyDoorOfTime = Actor_FindNearby(gPlayState, player, ACTOR_DOOR_TOKI, ACTORCAT_BG, 500.0f);
-        Actor* nearbyFrogs = Actor_FindNearby(gPlayState, player, ACTOR_EN_FR, ACTORCAT_NPC, 300.0f);
-        Actor* nearbyGossipStone = Actor_FindNearby(gPlayState, player, ACTOR_EN_GS, ACTORCAT_NPC, 300.0f);
-        bool justPlayedSoT = gPlayState->msgCtx.lastPlayedSong == OCARINA_SONG_TIME;
-        bool notNearAnySource = !nearbyTimeBlockEmpty && !nearbyTimeBlock && !nearbyOcarinaSpot && !nearbyDoorOfTime &&
-                                !nearbyFrogs && !nearbyGossipStone;
-        bool hasOcarinaOfTime = (INV_CONTENT(ITEM_OCARINA_TIME) == ITEM_OCARINA_TIME);
-        bool hasMasterSword = CHECK_OWNED_EQUIP(EQUIP_TYPE_SWORD, EQUIP_INV_SWORD_MASTER);
-        int timeTravelSetting = CVarGetInteger(CVAR_ENHANCEMENT("TimeTravel"), 0);
-        bool meetsTimeTravelRequirements = false;
-
-        switch (timeTravelSetting) {
-            case TIME_TRAVEL_ANY:
-                meetsTimeTravelRequirements = true;
-                break;
-            case TIME_TRAVEL_ANY_MS:
-                meetsTimeTravelRequirements = hasMasterSword;
-                break;
-            case TIME_TRAVEL_OOT_MS:
-                meetsTimeTravelRequirements = hasMasterSword && hasOcarinaOfTime;
-                break;
-            case TIME_TRAVEL_OOT:
-            default:
-                meetsTimeTravelRequirements = hasOcarinaOfTime;
-                break;
-        }
-
-        if (justPlayedSoT && notNearAnySource && meetsTimeTravelRequirements) {
-            SwitchAge();
-        }
-    });
-}
-
 static bool hasAffectedHealth = false;
 void UpdatePermanentHeartLossState() {
     if (!GameInteractor::IsSaveLoaded())
@@ -902,7 +807,6 @@ void InitMods() {
     RandomizerRegisterHooks();
     TimeSaverRegisterHooks();
     RegisterTTS();
-    RegisterOcarinaTimeTravel();
     RegisterPermanentHeartLoss();
     RegisterDeleteFileOnDeath();
     RegisterHyperBosses();
